% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{convertUnits}
\alias{convertUnits}
\title{convertUnits}
\usage{
convertUnits(value, from, to, factor = 1024)
}
\arguments{
\item{value}{value to convert}

\item{from}{units converting from}

\item{to}{units to convert to}

\item{factor}{conversion factor}
}
\description{
Convert between memory units.
}
\details{
Units can include B, KB, MB, GB
}
