% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAlive.R
\name{isAlive}
\alias{isAlive}
\title{Check if a host is active}
\usage{
isAlive(host, port = NULL)
}
\arguments{
\item{host}{host address}

\item{port}{port to check}
}
\description{
Check if a host is active.
}
\examples{
isAlive('www.google.com')
}
