% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memoryAvailable}
\alias{memoryAvailable}
\title{availableMemory}
\usage{
memoryAvailable(units = "GB")
}
\arguments{
\item{units}{units in which to return memory values (B,KB,MB,GB)}
}
\description{
Return available system memeory.
}
