% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memoryInfo}
\alias{memoryInfo}
\title{memory}
\usage{
memoryInfo(units = "KB")
}
\arguments{
\item{units}{units in which to return memory values (B,KB,MB,GB)}
}
\description{
Return memory usage information.
}
