% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memoryUsed}
\alias{memoryUsed}
\title{usedMemory}
\usage{
memoryUsed(units = "GB")
}
\arguments{
\item{units}{units in which to return memory values (B,KB,MB,GB)}
}
\description{
Return used system memeory.
}
