% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disks.R
\name{partitions}
\alias{partitions}
\title{partitions}
\usage{
partitions(units = "GB")
}
\arguments{
\item{units}{units in which to return memory values (B,KB,MB,GB)}
}
\description{
Summary of system disk partitions.
}
