% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{processes}
\alias{processes}
\title{processes}
\usage{
processes(root = TRUE)
}
\arguments{
\item{root}{should root processes be returned}
}
\description{
Return system processes
}
\examples{
processes()
}
