% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{totalMemory}
\alias{totalMemory}
\title{totalMemory}
\usage{
totalMemory(units = "GB")
}
\arguments{
\item{units}{units in which to return memory values (B,KB,MB,GB)}
}
\description{
Return total system memeory.
}
