% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptime.R
\name{uptime}
\alias{uptime}
\title{Current system uptime}
\usage{
uptime(unit = "second")
}
\arguments{
\item{unit}{a character string that specifies with time units to use (see \code{?lubridate::time_length})}
}
\description{
Return system uptime in seconds
}
