% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{userMemory}
\alias{userMemory}
\title{userMemory}
\usage{
userMemory(units = "GB")
}
\arguments{
\item{units}{units in which to return memory values (B,KB,MB,GB)}
}
\description{
Return user system memory usage.
}
