% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memory.R
\name{memoryUser}
\alias{memoryUser}
\title{System memory by user}
\usage{
memoryUser()
}
\arguments{
\item{units}{units in which to return memory values (B,KB,MB,GB)}
}
\description{
Return user system memory usage.
}
